+++ Dart Challange/unit_test_4/bin/unit_test_4.dart
@@ -1,0 +1,5 @@
+import 'package:unit_test_4/unit_test_4.dart' as unit_test_4;
+
+void main(List<String> arguments) {
+  unit_test_4.validatePassword(password: 'password');
+}
